requirements.txt
flask
python-telegram-bot==13.15
imageio
imageio-ffmpeg
numpy
scikit-image
PyYAML
torch
torchvision